more mysac work from upstream Thierry FOURNIER: fixes some memory allocation errors and includes some profiling functions
git-svn-id: http://svn.enlightenment.org/svn/e/trunk@58868 7cbeb6ba-43b4-40fd-8cce-4c39aea84d33

+++ PROTO/esskyuehl/src/lib/mysql/mysac/mysac_decode_field.c
@@ -45,7 +45,7 @@ int mysac_decode_field(char *buf, int len, MYSQL_FIELD *col) {
 	 2 (Length Coded Binary)    flags
 	 1                          decimals
 	 n (Length Coded Binary)    default
-
+	 
 	 -> VERSION 4.1
 	 Bytes                      Name
 	 -----                      ----
@@ -74,7 +74,7 @@ int mysac_decode_field(char *buf, int len, MYSQL_FIELD *col) {
 	if (tmp_len == -1)
 		return -MYERR_BAD_LCB;
 	i += tmp_len;
-	if ((unsigned long)(i + size) > (unsigned long)len)
+	if (i + size > (unsigned int)len)
 		return -MYERR_LEN_OVER_BUFFER;
 	col->catalog_length = size;
 	memmove(wh, &buf[i], size);
@@ -88,7 +88,7 @@ int mysac_decode_field(char *buf, int len, MYSQL_FIELD *col) {
 	if (tmp_len == -1)
 		return -MYERR_BAD_LCB;
 	i += tmp_len;
-	if ((unsigned long)(i + size) > (unsigned long)len)
+	if (i + size > (unsigned int)len)
 		return -MYERR_LEN_OVER_BUFFER;
 	col->db_length = size;
 	memmove(wh, &buf[i], size);
@@ -102,7 +102,7 @@ int mysac_decode_field(char *buf, int len, MYSQL_FIELD *col) {
 	if (tmp_len == -1)
 		return -MYERR_BAD_LCB;
 	i += tmp_len;
-	if ((unsigned long)(i + size) > (unsigned long)len)
+	if (i + size > (unsigned int)len)
 		return -MYERR_LEN_OVER_BUFFER;
 	col->table_length = size;
 	memmove(wh, &buf[i], size);
@@ -116,7 +116,7 @@ int mysac_decode_field(char *buf, int len, MYSQL_FIELD *col) {
 	if (tmp_len == -1)
 		return -MYERR_BAD_LCB;
 	i += tmp_len;
-	if ((unsigned long)(i + size) > (unsigned long)len)
+	if (i + size > (unsigned int)len)
 		return -MYERR_LEN_OVER_BUFFER;
 	col->org_table_length = size;
 	memmove(wh, &buf[i], size);
@@ -130,7 +130,7 @@ int mysac_decode_field(char *buf, int len, MYSQL_FIELD *col) {
 	if (tmp_len == -1)
 		return -MYERR_BAD_LCB;
 	i += tmp_len;
-	if ((unsigned long)(i + size) > (unsigned long)len)
+	if (i + size > (unsigned int)len)
 		return -MYERR_LEN_OVER_BUFFER;
 	col->name_length = size;
 	memmove(wh, &buf[i], size);
@@ -144,7 +144,7 @@ int mysac_decode_field(char *buf, int len, MYSQL_FIELD *col) {
 	if (tmp_len == -1)
 		return -MYERR_BAD_LCB;
 	i += tmp_len;
-	if ((unsigned long)(i + size) > (unsigned long)len)
+	if (i + size > (unsigned int)len)
 		return -MYERR_LEN_OVER_BUFFER;
 	col->org_name_length = size;
 	memmove(wh, &buf[i], size);
@@ -189,7 +189,7 @@ int mysac_decode_field(char *buf, int len, MYSQL_FIELD *col) {
 		if (tmp_len == -1)
 			return -MYERR_BAD_LCB;
 		i += tmp_len;
-		if ((unsigned long)(i + size) > (unsigned long)len)
+		if (i + size > (unsigned int)len)
 			return -MYERR_LEN_OVER_BUFFER;
 		col->def_length = size;
 		memmove(wh, &buf[i], size);
@@ -202,7 +202,7 @@ int mysac_decode_field(char *buf, int len, MYSQL_FIELD *col) {
 		col->def = NULL;
 		col->def_length = 0;
 	}
-
+		
 
 	/* set write pointer */
 	return wh - buf;
